refactor: simplify by combining index.js and msgs.js into primes.js
